Add files via upload
This Java program demonstrates the basic operations of a queue using the Queue interface and Linked List class. It adds elements (1 to 5) to the queue, displays all elements, uses the peek method to view the front element without removing it, employs the poll method to retrieve and remove the front element, and finally, the offer method to add a new element (9) to the queue. This showcases fundamental queue functionalities such as insertion, retrieval, and deletion.